Move CanisterMethod and CanisterMethodType
