+++ src/act/node/canister_method/canister_method_type.rs
@@ -1,0 +1,26 @@
+use std::fmt;
+
+#[derive(Clone, PartialEq)]
+pub enum CanisterMethodType {
+    Heartbeat,
+    Init,
+    InspectMessage,
+    PostUpgrade,
+    PreUpgrade,
+    Query,
+    Update,
+}
+
+impl fmt::Display for CanisterMethodType {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            CanisterMethodType::Heartbeat => write!(f, "Heartbeat"),
+            CanisterMethodType::Init => write!(f, "Init"),
+            CanisterMethodType::InspectMessage => write!(f, "InspectMessage"),
+            CanisterMethodType::PostUpgrade => write!(f, "PostUpgrade"),
+            CanisterMethodType::PreUpgrade => write!(f, "PreUpgrade"),
+            CanisterMethodType::Query => write!(f, "Query"),
+            CanisterMethodType::Update => write!(f, "Update"),
+        }
+    }
+}
